Add `isAnyCurrent()` getter in (resolved) `MenuItem`s + refactor tests
